Resolved C26451 errors where necessary (arithmetic overflow)
Resolved C26451 errors popping up in both the CPU and GPU timing result sections. This error is ignored in certain instances (e.g. container allocation).

It is also worth noting that some errors falsely flag arithmetic overflow (see https://stackoverflow.com/a/67278228) due to older VS builds.

Previous commit forgot to include definition for MASK_ONE_DIM in GPU oneConv; this is now resolved.

+++ CpuArithmeticApp/AppFunctions/src/twoConv/twoConvCore.cpp
@@ -38,7 +38,7 @@ void twoConvCore()
     twoConvCheck(mainVec, maskVec, resVec, conSize);
 
     // Calculate overall time spent to complete operation
-    double completionTime { (opEnd - opStart) / (double)CLOCKS_PER_SEC };
+    double completionTime{ ((static_cast<double>(opEnd)) - (static_cast<double>(opStart))) / (double)CLOCKS_PER_SEC };
 
     // Output timing to complete operation and container size
     std::cout << completionTime << "s 2D Convolution computation time, with a container size of " << conSize * conSize << ".\n\n";
