+++ GpuArithmeticApp/AppFunctions/src/matMulti/matMultiCore.cu
@@ -70,7 +70,7 @@ void matMultiCore()
 	cudaFree(deviceC);
 
 	// Calculate overall time spent to complete operation
-	double completionTime{ (opEnd - opStart) / (double)CLOCKS_PER_SEC };
+	double completionTime{ ((static_cast<double>(opEnd)) - (static_cast<double>(opStart))) / (double)CLOCKS_PER_SEC };
 
 	// Output timing to complete operation and container size
 	std::cout << completionTime << "s Matrix Multiplication computation time, with a container size of " << conSize * conSize << ".\n\n";
