+++ GpuArithmeticApp/AppFunctions/src/twoConv/twoConvCore.cu
@@ -81,7 +81,7 @@ void twoConvCore()
     cudaFree(deviceResVec);
 
     // Calculate overall time spent to complete operation
-    double completionTime{ (opEnd - opStart) / (double)CLOCKS_PER_SEC };
+    double completionTime{ ((static_cast<double>(opEnd)) - (static_cast<double>(opStart))) / (double)CLOCKS_PER_SEC };
 
     // Output timing to complete operation and container size
     std::cout << completionTime << "s 2D Convolution computation time, with a container size of " << conSize * conSize << ".\n\n";
